dist/datachannel_test_2peer/send/005.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Send Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
            var label = "test";
            var dataChannel = null;
            var expected = "blob";
            var myBlobData = "<strong> This is a blob </strong><br /> Testing send a blob with webRTC";
	        var myBlob = new Blob([myBlobData], {
	            type : "text/plain"
	        });
            var test = async_test("Set up a DataChannel - send and receive a Blob - check type and received data");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    dataChannel.onopen = test.step_func(function() {
                        dataChannel.send(myBlob);
                    });
                    dataChannel.onmessage = test.step_func(function(e) {
						assert_true(e.data instanceof Blob, "No instance of blob ");
			            assert_equals(e.data.size, myBlob.size, "Received wrong size ");
			            // read content from a Blob with a FileReader
			            var reader = new FileReader();
			            reader.addEventListener("loadend", function() {
			                assert_equals(reader.result, myBlobData, "Wrong Data ");
			                test.done();
			            });
			            reader.readAsText(e.data);
                    });
                });
            } else {
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    dataChannel.onmessage = test.step_func(function(e) {
						assert_true(e.data instanceof Blob, "No instance of blob ");
			            assert_equals(e.data.size, myBlob.size, "Received wrong size ");
			            // read content from a Blob with a FileReader
			            var reader = new FileReader();
			            reader.addEventListener("loadend", function() {
			                assert_equals(reader.result, myBlobData, "Wrong Data ");
			                dataChannel.send(myBlob);
			                test.done();
			            });
			            reader.readAsText(e.data);
                    });
                });
            }
		</script>
	</body>
</html>

